feat: move api folder to lib n update imports
