+++ app/api/posts/user/[id]/route.ts
@@ -1,7 +1,7 @@
-import { createErrorResponse } from '@api/createErrorResponse';
-import { filterPosts } from '@api/filterPosts';
-import { getPaginationParams } from '@api/getPaginationParams';
-import { paginateData } from '@api/paginateData';
+import { createErrorResponse } from '@/lib/api/createErrorResponse';
+import { filterPosts } from '@/lib/api/filterPosts';
+import { getPaginationParams } from '@/lib/api/getPaginationParams';
+import { paginateData } from '@/lib/api/paginateData';
 import { NOT_FOUND_ERROR_STATUS_CODE, USERS_LIMIT_DEFAULT } from '@constants/api';
 import posts from '@lib/mocks/posts';
