+++ app/api/users/[id]/route.ts
@@ -1,5 +1,5 @@
-import { filterPosts } from '@/api/filterPosts';
-import { createErrorResponse } from '@api/createErrorResponse';
+import { createErrorResponse } from '@/lib/api/createErrorResponse';
+import { filterPosts } from '@/lib/api/filterPosts';
 import { NOT_FOUND_ERROR_STATUS_CODE } from '@constants/api';
 import posts from '@lib/mocks/posts';
 import users from '@lib/mocks/users';
