+++ app/api/users/route.ts
@@ -1,5 +1,5 @@
-import { getPaginationParams } from '@api/getPaginationParams';
-import { paginateData } from '@api/paginateData';
+import { getPaginationParams } from '@/lib/api/getPaginationParams';
+import { paginateData } from '@/lib/api/paginateData';
 import { USERS_LIMIT_DEFAULT } from '@constants/api';
 import users from '@lib/mocks/users';
