+++ hooks/__tests__/useGetUsers.test.tsx
@@ -1,7 +1,7 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { render } from '@testing-library/react';
 
-import { getAllUsers } from '@api/getAllUsers';
+import { getAllUsers } from '@/lib/api/getAllUsers';
 import { useGetUsers } from '@hooks/useUsers';
 import { User, UsersResponse } from '@lib/zod/user';
