+++ lib/api/__tests__/createErrorResponse.test.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 
-import { createErrorResponse } from '@api/createErrorResponse';
+import { createErrorResponse } from '@/lib/api/createErrorResponse';
 
 jest.mock('next/server', () => ({
   NextResponse: {
