+++ lib/api/__tests__/paginateData.test.ts
@@ -1,4 +1,4 @@
-import { paginateData } from '@api/paginateData';
+import { paginateData } from '@/lib/api/paginateData';
 
 describe('paginateData', () => {
   const data = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
