+++ lib/api/createErrorResponse.ts
