+++ middleware.ts
@@ -10,5 +10,5 @@ export default function middleware(request: NextRequest) {
 }
 
 export const config = {
-  matcher: ['/((?!api|_next/static|_next/image|images|favicon.ico).*)'],
+  matcher: ['/', '/((?!api|_next/static|_next/image|images|favicon.ico).*)'],
 };
