+++ next.config.js
@@ -9,7 +9,6 @@ const nextConfig = {
   images: {
     domains: ['s3-alpha-sig.figma.com'],
   },
-  trailingSlash: true,
 };
 
 export default withNextIntl(nextConfig);
